fix(misc): minor bug fixes, test updates, and tilde expansion refinements
ContextMenu positioning, SuggestOverlay styling, UpdateProgressDialog
error handling, useRepository cleanup, terminals store fixes. Update
tests for DnD refactor, plugin registry, PluginsTab sort, and
claude-wakeup plugin. Refine tilde expansion in cli/pty/stdio_client.

+++ src-tauri/src/mcp_proxy/stdio_client.rs
@@ -394,7 +394,7 @@ impl StdioMcpClient {
         for _ in 0..MAX_SKIP {
             let msg = self.read_line()?;
             match msg.get("id") {
-                Some(resp_id) if resp_id.as_u64() == Some(id as u64) => return Ok(msg),
+                Some(resp_id) if resp_id.as_u64() == Some(id) => return Ok(msg),
                 Some(_) => {
                     tracing::debug!(
                         upstream = %self.config.name,
